backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Configuration
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis + Celery (task queue)
redis==5.0.4
celery==5.4.0

# HTTP client (for OpenAI, Anthropic, Cohere)
httpx==0.27.0

# AI / ML
openai==1.30.1
anthropic==0.28.0
cohere==5.5.3
weaviate-client==4.6.1
tiktoken==0.7.0

# Document parsing
pymupdf==1.24.3
python-docx==1.1.2
python-magic==0.4.27

# AWS
boto3==1.34.101

# Utilities
python-multipart==0.0.9
python-dotenv==1.0.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
